+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/dcql/DCQLCredentialQueryResponse.kt
@@ -1,8 +1,0 @@
-package at.asitplus.openid.dcql
-
-import kotlinx.serialization.json.JsonElement
-
-data class DCQLCredentialQueryResponse(
-    val jsonElement: JsonElement,
-)
-