+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/dcql/DCQLQueryResponse.kt
@@ -1,6 +1,6 @@
 package at.asitplus.openid.dcql
 
-data class DCQLQueryResponse(
+data class DCQLQueryResponse<DCQLCredentialQueryResponse: Any>(
     val submissions: Map<DCQLCredentialQueryIdentifier, List<DCQLCredentialQueryResponse>>,
 )